Try replacing these files for t31 4.4 ISP and building from SDK
