+++ 4.4/isp/t31/Kbuild
@@ -1,6 +1,5 @@
 SRCS := $(DIR)/tx-isp-debug.c \
-	$(DIR)/tx-isp-module.c \
-	$(DIR)/tx-isp-funcs.c
+	$(DIR)/tx-isp-module.c
 
 OBJS := $(SRCS:%.c=%.o) $(ASM_SRCS:%.S=%.o)
 OBJS += $(KERNEL_VERSION)/sdk/lib$(SOC_FAMILY)-firmware.a
