+++ 4.4/isp/t31/include/tx-isp-common.h
@@ -7,7 +7,6 @@
 #include <linux/spi/spi.h>
 #include <linux/videodev2.h>
 #include <linux/v4l2-mediabus.h>
-#include <linux/media-bus-format.h>
 #include <media/media-entity.h>
 #include <media/media-device.h>
 #include <media/v4l2-subdev.h>
@@ -29,9 +28,6 @@
 #define TX_ISP_DS1_CAHNNEL_MAX_WIDTH		1920
 #define TX_ISP_DS1_CAHNNEL_MAX_HEIGHT		1080
 
-/* Color settings */
-#define V4L2_MBUS_FMT_SBGGR10_1X10 0x3007
-#define V4L2_MBUS_FMT_SRGGB10_1X10 0x300f
 
 /*****************************************************
  			sensor attributes
@@ -117,11 +113,87 @@ enum rgbir_mbus_fmt{
 	V4L2_MBUS_FMT_SIGGB12_1X12,
 };
 
+enum v4l2_mbus_pixelcode {
+	V4L2_MBUS_FMT_FIXED = 0x0001,
+
+	/* RGB - next is 0x100d */
+	V4L2_MBUS_FMT_RGB444_2X8_PADHI_BE = 0x1001,
+	V4L2_MBUS_FMT_RGB444_2X8_PADHI_LE = 0x1002,
+	V4L2_MBUS_FMT_RGB555_2X8_PADHI_BE = 0x1003,
+	V4L2_MBUS_FMT_RGB555_2X8_PADHI_LE = 0x1004,
+	V4L2_MBUS_FMT_BGR565_2X8_BE = 0x1005,
+	V4L2_MBUS_FMT_BGR565_2X8_LE = 0x1006,
+	V4L2_MBUS_FMT_RGB565_2X8_BE = 0x1007,
+	V4L2_MBUS_FMT_RGB565_2X8_LE = 0x1008,
+	V4L2_MBUS_FMT_RGB666_1X18 = 0x1009,
+	V4L2_MBUS_FMT_RGB888_1X24 = 0x100a,
+	V4L2_MBUS_FMT_RGB888_2X12_BE = 0x100b,
+	V4L2_MBUS_FMT_RGB888_2X12_LE = 0x100c,
+
+	/* YUV (including grey) - next is 0x2017 */
+	V4L2_MBUS_FMT_Y8_1X8 = 0x2001,
+	V4L2_MBUS_FMT_UV8_1X8 = 0x2015,
+	V4L2_MBUS_FMT_UYVY8_1_5X8 = 0x2002,
+	V4L2_MBUS_FMT_VYUY8_1_5X8 = 0x2003,
+	V4L2_MBUS_FMT_YUYV8_1_5X8 = 0x2004,
+	V4L2_MBUS_FMT_YVYU8_1_5X8 = 0x2005,
+	V4L2_MBUS_FMT_UYVY8_2X8 = 0x2006,
+	V4L2_MBUS_FMT_VYUY8_2X8 = 0x2007,
+	V4L2_MBUS_FMT_YUYV8_2X8 = 0x2008,
+	V4L2_MBUS_FMT_YVYU8_2X8 = 0x2009,
+	V4L2_MBUS_FMT_Y10_1X10 = 0x200a,
+	V4L2_MBUS_FMT_YUYV10_2X10 = 0x200b,
+	V4L2_MBUS_FMT_YVYU10_2X10 = 0x200c,
+	V4L2_MBUS_FMT_Y12_1X12 = 0x2013,
+	V4L2_MBUS_FMT_UYVY8_1X16 = 0x200f,
+	V4L2_MBUS_FMT_VYUY8_1X16 = 0x2010,
+	V4L2_MBUS_FMT_YUYV8_1X16 = 0x2011,
+	V4L2_MBUS_FMT_YVYU8_1X16 = 0x2012,
+	V4L2_MBUS_FMT_YDYUYDYV8_1X16 = 0x2014,
+	V4L2_MBUS_FMT_YUYV10_1X20 = 0x200d,
+	V4L2_MBUS_FMT_YVYU10_1X20 = 0x200e,
+	V4L2_MBUS_FMT_YUV10_1X30 = 0x2016,
+
+	/* Bayer - next is 0x3019 */
+	V4L2_MBUS_FMT_SBGGR8_1X8 = 0x3001,
+	V4L2_MBUS_FMT_SGBRG8_1X8 = 0x3013,
+	V4L2_MBUS_FMT_SGRBG8_1X8 = 0x3002,
+	V4L2_MBUS_FMT_SRGGB8_1X8 = 0x3014,
+	V4L2_MBUS_FMT_SBGGR10_ALAW8_1X8 = 0x3015,
+	V4L2_MBUS_FMT_SGBRG10_ALAW8_1X8 = 0x3016,
+	V4L2_MBUS_FMT_SGRBG10_ALAW8_1X8 = 0x3017,
+	V4L2_MBUS_FMT_SRGGB10_ALAW8_1X8 = 0x3018,
+	V4L2_MBUS_FMT_SBGGR10_DPCM8_1X8 = 0x300b,
+	V4L2_MBUS_FMT_SGBRG10_DPCM8_1X8 = 0x300c,
+	V4L2_MBUS_FMT_SGRBG10_DPCM8_1X8 = 0x3009,
+	V4L2_MBUS_FMT_SRGGB10_DPCM8_1X8 = 0x300d,
+	V4L2_MBUS_FMT_SBGGR10_2X8_PADHI_BE = 0x3003,
+	V4L2_MBUS_FMT_SBGGR10_2X8_PADHI_LE = 0x3004,
+	V4L2_MBUS_FMT_SBGGR10_2X8_PADLO_BE = 0x3005,
+	V4L2_MBUS_FMT_SBGGR10_2X8_PADLO_LE = 0x3006,
+	V4L2_MBUS_FMT_SBGGR10_1X10 = 0x3007,
+	V4L2_MBUS_FMT_SGBRG10_1X10 = 0x300e,
+	V4L2_MBUS_FMT_SGRBG10_1X10 = 0x300a,
+	V4L2_MBUS_FMT_SRGGB10_1X10 = 0x300f,
+	V4L2_MBUS_FMT_SBGGR12_1X12 = 0x3008,
+	V4L2_MBUS_FMT_SGBRG12_1X12 = 0x3010,
+	V4L2_MBUS_FMT_SGRBG12_1X12 = 0x3011,
+	V4L2_MBUS_FMT_SRGGB12_1X12 = 0x3012,
+
+	/* JPEG compressed formats - next is 0x4002 */
+	V4L2_MBUS_FMT_JPEG_1X8 = 0x4001,
+
+	/* Vendor specific formats - next is 0x5002 */
+
+	/* S5C73M3 sensor specific interleaved UYVY and JPEG */
+	V4L2_MBUS_FMT_S5C_UYVY_JPEG_1X8 = 0x5001,
+};
+
 struct tx_isp_sensor_win_setting {
 	int	width;
 	int	height;
 	int fps;
-	int mbus_code;
+	enum v4l2_mbus_pixelcode mbus_code;
 	enum v4l2_colorspace colorspace;
 	void *regs;	/* Regs to tweak; the default fps is fast */
 };
@@ -188,7 +260,6 @@ enum tx_sensor_csi_fmt {
 	TX_SENSOR_RAW8 = 0,
 	TX_SENSOR_RAW10,
 	TX_SENSOR_RAW12,
-	TX_SENSOR_YUV422 = 7, /**> This mode requires the use of specific sensor drivers(such as TP2850) and bin files*/
 };
 
 typedef enum {
@@ -460,8 +531,6 @@ enum tx_isp_notification {
 	TX_ISP_EVENT_SENSOR_WDR_OPEN,
 	TX_ISP_EVENT_SENSOR_LOGIC,
 	TX_ISP_EVENT_SENSOR_EXPO,
-    TX_ISP_EVENT_GPIO_INIT,
-    TX_ISP_EVENT_SET_GPIO_STATE,
 	/* the events of frame-channel are defined as follows. */
 	TX_ISP_EVENT_FRAME_CHAN_BYPASS_ISP = NOTIFICATION_TYPE_FS_OPS,
 	TX_ISP_EVENT_FRAME_CHAN_GET_FMT,
@@ -545,21 +614,18 @@ struct isp_buf_info {
 	uint32_t paddr;
 	uint32_t size;
 };
-
-struct isp_set_gpio_attr{
-    uint16_t gpio_num[10];
-    uint16_t gpio_sta[10];
-    uint16_t free;
-};
+#define BASE_DEVICE_PRIVATE         0
+//#define BASE_VIDIOC_PRIVATE 192
 
 #define VIDIOC_REGISTER_SENSOR			  _IOW('V', BASE_VIDIOC_PRIVATE + 1, struct tx_isp_sensor_register_info)
 #define VIDIOC_RELEASE_SENSOR			  _IOW('V', BASE_VIDIOC_PRIVATE + 2, struct tx_isp_sensor_register_info)
 #define VIDIOC_SET_FRAME_FORMAT			  _IOWR('V', BASE_VIDIOC_PRIVATE + 3, struct frame_image_format)
 #define VIDIOC_GET_FRAME_FORMAT               _IOR('V', BASE_VIDIOC_PRIVATE + 4, struct frame_image_format)
+//#define VIDIOC_SET_FRAME_FORMAT			  _IOWR('V', BASE_VIDIOC_PRIVATE + 3, int)
+//#define VIDIOC_GET_FRAME_FORMAT               _IOR('V', BASE_VIDIOC_PRIVATE + 4, int)
+
 #define VIDIOC_DEFAULT_CMD_SET_BANKS	      _IOW('V', BASE_VIDIOC_PRIVATE + 5, int)
 #define VIDIOC_DEFAULT_CMD_ISP_TUNING	      _IOWR('V', BASE_VIDIOC_PRIVATE + 6, struct isp_image_tuning_default_ctrl)
-#define VIDIOC_SET_DEFAULT_BIN_PATH               _IOWR('V',BASE_VIDIOC_PRIVATE + 7, int)
-#define VIDIOC_GET_DEFAULT_BIN_PATH               _IOWR('V',BASE_VIDIOC_PRIVATE + 8, int)
 
 #define VIDIOC_CREATE_SUBDEV_LINKS            _IOW('V', BASE_VIDIOC_PRIVATE + 16, int)
 #define VIDIOC_DESTROY_SUBDEV_LINKS	 		  _IOW('V', BASE_VIDIOC_PRIVATE + 17, int)
@@ -584,11 +650,36 @@ struct isp_set_gpio_attr{
 #define TISP_VIDIOC_SET_AWB_ALGO_OPEN				_IOWR('V',BASE_VIDIOC_PRIVATE + 35, int)
 #define TISP_VIDIOC_SET_AWB_ALGO_CLOSE				_IOWR('V',BASE_VIDIOC_PRIVATE + 36, int)
 #define TISP_VIDIOC_SET_AWB_ALGO_CTRL				_IOWR('V',BASE_VIDIOC_PRIVATE + 37, int)
-#define TISP_VIDIOC_SET_FRAME_DROP				_IOWR('V',BASE_VIDIOC_PRIVATE + 38, int)
-#define TISP_VIDIOC_GET_FRAME_DROP				_IOWR('V',BASE_VIDIOC_PRIVATE + 39, int)
 
-#define VIDIOC_SET_GPIO_INIT                              _IOWR('V',BASE_VIDIOC_PRIVATE + 40, int)
-#define VIDIOC_SET_GPIO_STA                              _IOWR('V',BASE_VIDIOC_PRIVATE + 41, int)
+//add
+#define VIDIOC_DEFAULT_CMD_LISTEN_BUF    _IOR('V', BASE_VIDIOC_PRIVATE - 1, int)
+
+#define TISP_VIDIOC_DRIVER_VERSION              _IOW('T', BASE_DEVICE_PRIVATE + 1, int)
+#define TISP_VIDIOC_ENUMINPUT                   _IOWR('T', BASE_DEVICE_PRIVATE + 2, int)
+#define TISP_VIDIOC_G_INPUT                     _IOW('T', BASE_DEVICE_PRIVATE + 3, struct tx_isp_initarg)
+#define TISP_VIDIOC_S_INPUT                 _IOWR('T', BASE_DEVICE_PRIVATE + 4, struct tx_isp_initarg)
+#define TISP_VIDIOC_REGISTER_SENSOR             _IOW('T', BASE_DEVICE_PRIVATE + 5, struct tx_isp_sensor_register_info)
+#define TISP_VIDIOC_RELEASE_SENSOR              _IOW('T', BASE_DEVICE_PRIVATE + 6, struct tx_isp_sensor_register_info)
+#define TISP_VIDIOC_STREAMON                    _IOW('T', BASE_DEVICE_PRIVATE + 7, struct tx_isp_initarg)
+#define TISP_VIDIOC_STREAMOFF                   _IOW('T', BASE_DEVICE_PRIVATE + 8, struct tx_isp_initarg)
+#define TISP_VIDIOC_CREATE_SUBDEV_LINKS             _IOW('T', BASE_DEVICE_PRIVATE + 9, struct tx_isp_initarg)
+#define TISP_VIDIOC_DESTROY_SUBDEV_LINKS            _IOW('T', BASE_DEVICE_PRIVATE + 10, struct tx_isp_initarg)
+#define TISP_VIDIOC_LINKS_STREAMON              _IOWR('T',BASE_DEVICE_PRIVATE + 11, struct tx_isp_initarg)
+#define TISP_VIDIOC_LINKS_STREAMOFF             _IOWR('T',BASE_DEVICE_PRIVATE + 12, struct tx_isp_initarg)
+#define TISP_VIDIOC_SET_SENSOR_REGISTER             _IOWR('T',BASE_DEVICE_PRIVATE + 13, struct tx_isp_dbg_register)
+#define TISP_VIDIOC_GET_SENSOR_REGISTER             _IOW('T', BASE_DEVICE_PRIVATE + 14, struct tx_isp_dbg_register)
+#define TISP_VIDIOC_SET_MDNS_BUF_INFO               _IOW('T', BASE_DEVICE_PRIVATE + 15, struct isp_buf_info)
+#define TISP_VIDIOC_GET_MDNS_BUF_INFO               _IOW('T', BASE_DEVICE_PRIVATE + 16, struct isp_buf_info)
+#define TISP_VIDIOC_SET_WDR_BUF_INFO                _IOW('T', BASE_DEVICE_PRIVATE + 17, struct isp_buf_info)
+#define TISP_VIDIOC_GET_WDR_BUF_INFO                _IOW('T', BASE_DEVICE_PRIVATE + 18, struct isp_buf_info)
+#define TISP_VIDIOC_ISP_WDR_ENABLE              _IOW('T', BASE_DEVICE_PRIVATE + 19, int)
+#define TISP_VIDIOC_ISP_WDR_DISABLE             _IOW('T', BASE_DEVICE_PRIVATE + 20, int)
+#define TISP_VIDIOC_SET_DUALSENSOR_MODE             _IOWR('T',BASE_DEVICE_PRIVATE + 21, int)
+#define TISP_VIDIOC_SET_DUALSENSOR_BUF_INFO         _IOWR('T',BASE_DEVICE_PRIVATE + 22, struct isp_buf_info)
+#define TISP_VIDIOC_GET_DUALSENSOR_BUF_INFO         _IOW('T', BASE_DEVICE_PRIVATE + 23, struct isp_buf_info)
+
+#define TISP_VIDIOC_BYPASS_SENSOR                               _IOW('T', BASE_DEVICE_PRIVATE + 24, struct tx_isp_initarg)
+#define TISP_VIDIOC_GET_AE_ALGO_HANDLE_SEC                            _IOWR('T',BASE_DEVICE_PRIVATE + 32, int)
 
 enum tx_isp_vidioc_default_command {
 	TX_ISP_VIDIOC_DEFAULT_CMD_BYPASS_ISP,
