+++ 4.4/isp/t31/include/tx-isp-debug.h
@@ -38,5 +38,4 @@ bool private_schedule_work(struct work_struct *work);
 void private_do_gettimeofday(struct timeval *tv);
 void private_dma_sync_single_for_device(struct device *dev,
 							      dma_addr_t addr, size_t size, enum dma_data_direction dir);
-__must_check int private_get_driver_interface(struct jz_driver_common_interfaces **pfaces);
 #endif /* _ISP_DEBUG_H_ */
