+++ 4.4/isp/t31/include/tx-isp-device.h
@@ -249,8 +249,8 @@ struct tx_isp_device {
 int private_reset_tx_isp_module(enum tx_isp_subdev_id id);
 int tx_isp_reg_set(struct tx_isp_subdev *sd, unsigned int reg, int start, int end, int val);
 
-extern int tx_isp_subdev_init(struct platform_device *pdev, struct tx_isp_subdev *sd, struct tx_isp_subdev_ops *ops);
-extern void tx_isp_subdev_deinit(struct tx_isp_subdev *sd);
+int tx_isp_subdev_init(struct platform_device *pdev, struct tx_isp_subdev *sd, struct tx_isp_subdev_ops *ops);
+void tx_isp_subdev_deinit(struct tx_isp_subdev *sd);
 int tx_isp_send_event_to_remote(struct tx_isp_subdev_pad *pad, unsigned int cmd, void *data);
 
 static inline void tx_isp_set_module_nodeops(struct tx_isp_module *module, struct file_operations *ops)
